Make PyInstaller generate a manifest file for the final executable and modify the bootloader to procedurally activate it at startup.
This fixes a problem with dynamically-loading with ctypes a module
that is bundled as assembly with PyInstaller, such as msvcrt; some
modules of the standard library do this (eg: uuid) and in turn this
causes problems with third-party libraries.

Patch mostly by Florian, I've just cleaned up the bootloader part.


git-svn-id: http://svn.pyinstaller.org/trunk@903 8dd32b29-ccff-0310-8a9a-9233e24343b1